refactoring file structure -- everything should be working in regards to selectedNode/displayOptions/edgeState rerendering
